perf(web): optimización de imágenes, JS y robustez tras auditoría completa
Auditoría de rendimiento y profesionalidad (5 pasadas + Lighthouse). El
sitio partía bien (98/95/100/90, CLS 0) pero con problemas concretos.

Imágenes (la queja original: "las fotos tardan o cargan según scroll"):
- La captura de cada ficha —el LCP de la página— iba con loading="lazy";
  ahora lleva priority y se precarga desde el head.
- Las primeras tarjetas de "IAs Destacadas" y "Nuevas adiciones", visibles
  sin scroll, forzaban lazy; ahora las 4 primeras van con priority.
- Logos de comparativas con priority.
- /images y /logos se servían con max-age=0 (revalidación en cada visita);
  ahora immutable de 1 año.
- 6 logos de origen desproporcionado (hasta 1,5 MB para iconos de 48 px)
  redimensionados a 256 px: 4,6 MB → 260 KB.

JavaScript y TTFB:
- /articles/[slug] era un componente cliente que embarcaba articles.ts
  entero (el blog completo) en cada visita. Ahora la página es servidor,
  resuelve el artículo y pasa solo ese artículo a ArticleView (cliente).
  First Load JS: 171 → 132 kB, y la ruta pasa a SSG. De paso, un slug
  inexistente devuelve 404 real (antes: 200 con shell, soft-404).
- La home leía searchParams en servidor, lo que la volvía dinámica y sin
  caché de edge (TTFB ~240 ms vs ~120). Ahora es estática: HomeContainer
  lee ?q= y ?category= de la URL tras hidratar, que es cuando se usaban.
- Speculation Rules: prerender especulativo de fichas, comparativas y
  categorías al pasar el ratón.

Robustez y cumplimiento:
- not-found.tsx y error.tsx globales con marca (solo existían para
  categorías).
- DataFast se cargaba con script inline saltándose el consentimiento de
  cookies que sí respetan GA y Clarity; ahora pasa por el mismo gate.
- Estado vacío en la búsqueda de /herramientas.
- Página /sobre con autoría y criterio editorial, enlazada en el footer y
  en el sitemap.

Accesibilidad:
- 13 secciones de la home tenían tarjetas div+onClick inaccesibles con
  teclado e invisibles para Google como enlaces; ahora son <a> reales con
  rel="noopener noreferrer" (CategoriesGrid, de estado interno, lleva
  role/tabIndex/teclado).
- Foco visible global con :focus-visible (cubre los 30+ outline-none) y
  prefers-reduced-motion respetado.
- aria-label en los inputs que solo tenían placeholder.

Limpieza:
- Preconnects muertos fuera (Google Fonts que next/font ya no usa, y los
  subdominios fastly de Mux que cambian por región).
- fontFamily 'Inter' fantasma eliminada de ~20 componentes (la fuente no
  se carga; forzaba la del sistema en vez de heredar Geist).
- Lighthouse del CI apuntaba a aifinder-five.vercel.app sin umbrales
  (pasaba siempre); ahora mide www.aifinder.es con mínimos exigidos.

Verificación: tsc limpio, 0 errores lint, build 396 páginas, 270/270
tests, y en servidor local: home estática conservando el H1 SEO, 404
reales, /sobre operativa, captura de ficha precargada desde el head.
