+++ .github/workflows/ci.yml
@@ -60,6 +60,7 @@ jobs:
       - name: Run Lighthouse CI
         uses: treosh/lighthouse-ci-action@v11
         with:
-          urls: 'https://aifinder-five.vercel.app/'
+          urls: 'https://www.aifinder.es/'
+          configPath: '.lighthouserc.json'
           uploadArtifacts: true
           temporaryPublicStorage: true
