+++ src/app/globals.css
@@ -55,3 +55,23 @@ body {
   max-width: 100vw;
   overflow-x: hidden;
 }
+
+/* Foco visible con teclado en todo el sitio: hay 30+ usos de outline-none
+   sin sustituto, y esta regla los cubre de una vez sin afectar al clic
+   (solo se activa con :focus-visible, no con :focus de ratón). */
+*:focus-visible {
+  outline: 2px solid #3b82f6 !important;
+  outline-offset: 2px;
+}
+
+/* Respetar la preferencia de menos movimiento del sistema */
+@media (prefers-reduced-motion: reduce) {
+  *,
+  *::before,
+  *::after {
+    animation-duration: 0.01ms !important;
+    animation-iteration-count: 1 !important;
+    transition-duration: 0.01ms !important;
+    scroll-behavior: auto !important;
+  }
+}
