+++ src/components/Sections/AIAutomationChatbotsSection.tsx
@@ -90,10 +90,6 @@ export default function AIAutomationChatbotsSection({
     },
   ];
 
-  const handleToolClick = (url: string) => {
-    window.open(url, '_blank');
-  };
-
   return (
     <div className="py-4 px-4 max-w-7xl mx-auto">
       {/* Vista móvil: Lista vertical con iconos */}
@@ -101,20 +97,19 @@ export default function AIAutomationChatbotsSection({
         {/* Contenedor Automatización - Móvil */}
         <div className="bg-zinc-950 rounded-xl p-4 border border-zinc-800">
           <div className="mb-6">
-            <h2
-              className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-              style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-            >
+            <h2 className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
               Automatización
             </h2>
           </div>
 
           <div className="space-y-3">
             {automationTools.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="flex items-start gap-3 py-1 pl-1 cursor-pointer hover:bg-zinc-900/50 rounded-lg transition-colors p-2"
-                onClick={() => handleToolClick(tool.url)}
               >
                 <Image
                   src={tool.logo}
@@ -128,7 +123,7 @@ export default function AIAutomationChatbotsSection({
                   <h3 className="font-bold text-white text-base mb-1">{tool.name}</h3>
                   <p className="text-zinc-400 text-sm leading-relaxed">{tool.description}</p>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
@@ -136,20 +131,19 @@ export default function AIAutomationChatbotsSection({
         {/* Contenedor Chatbots - Móvil */}
         <div className="bg-zinc-950 rounded-xl p-4 border border-zinc-800">
           <div className="mb-6">
-            <h2
-              className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-              style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-            >
+            <h2 className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
               Chatbots Empresariales
             </h2>
           </div>
 
           <div className="space-y-3">
             {chatbotTools.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="flex items-start gap-3 py-1 pl-1 cursor-pointer hover:bg-zinc-900/50 rounded-lg transition-colors p-2"
-                onClick={() => handleToolClick(tool.url)}
               >
                 <Image
                   src={tool.logo}
@@ -163,7 +157,7 @@ export default function AIAutomationChatbotsSection({
                   <h3 className="font-bold text-white text-base mb-1">{tool.name}</h3>
                   <p className="text-zinc-400 text-sm leading-relaxed">{tool.description}</p>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
@@ -174,10 +168,7 @@ export default function AIAutomationChatbotsSection({
         {/* Contenedor Automatización */}
         <div className="bg-zinc-950 rounded-xl p-3 border border-zinc-800">
           <div className="mb-4">
-            <h2
-              className="text-lg md:text-xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-              style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-            >
+            <h2 className="text-lg md:text-xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
               Automatización
             </h2>
           </div>
@@ -185,10 +176,12 @@ export default function AIAutomationChatbotsSection({
           {/* Grid de 2 columnas para herramientas de automatización */}
           <div className="grid grid-cols-2 gap-2">
             {automationTools.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="group cursor-pointer p-2 hover:bg-zinc-900/30 rounded-lg transition-all duration-200"
-                onClick={() => handleToolClick(tool.url)}
               >
                 <div className="flex items-center gap-3">
                   {/* Logo */}
@@ -211,7 +204,7 @@ export default function AIAutomationChatbotsSection({
                     </p>
                   </div>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
@@ -219,10 +212,7 @@ export default function AIAutomationChatbotsSection({
         {/* Contenedor Chatbots */}
         <div className="bg-zinc-950 rounded-xl p-3 border border-zinc-800">
           <div className="mb-4">
-            <h2
-              className="text-lg md:text-xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-              style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-            >
+            <h2 className="text-lg md:text-xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
               Chatbots Empresariales
             </h2>
           </div>
@@ -230,10 +220,12 @@ export default function AIAutomationChatbotsSection({
           {/* Grid de 2 columnas para herramientas de chatbots */}
           <div className="grid grid-cols-2 gap-2">
             {chatbotTools.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="group cursor-pointer p-2 hover:bg-zinc-900/30 rounded-lg transition-all duration-200"
-                onClick={() => handleToolClick(tool.url)}
               >
                 <div className="flex items-center gap-3">
                   {/* Logo */}
@@ -256,7 +248,7 @@ export default function AIAutomationChatbotsSection({
                     </p>
                   </div>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
