+++ src/components/Sections/AIRecommendationsSection.tsx
@@ -77,20 +77,13 @@ export default function AIRecommendationsSection({
     },
   ];
 
-  const handleRecommendationClick = (url: string) => {
-    window.open(url, '_blank');
-  };
-
   return (
     <div className="py-4 px-4 max-w-7xl mx-auto">
       {/* Contenedor principal con bordes */}
       <div className="bg-zinc-950 rounded-xl p-4 border border-zinc-800">
         {/* Header con título */}
         <div className="flex items-center justify-between mb-8">
-          <h2
-            className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-            style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-          >
+          <h2 className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
             Recomendaciones IA
           </h2>
         </div>
@@ -99,10 +92,12 @@ export default function AIRecommendationsSection({
         <div className="md:hidden">
           <div className="space-y-6">
             {recommendations.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="group cursor-pointer"
-                onClick={() => handleRecommendationClick(tool.url)}
               >
                 {/* Imagen de la herramienta */}
                 <div className="relative aspect-video bg-zinc-800 rounded-lg mb-3 overflow-hidden border border-[#232323]">
@@ -145,7 +140,7 @@ export default function AIRecommendationsSection({
                     <span>{tool.description}</span>
                   </div>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
@@ -153,10 +148,12 @@ export default function AIRecommendationsSection({
         {/* Vista desktop: Grid de tarjetas */}
         <div className="hidden md:grid md:grid-cols-4 gap-6">
           {recommendations.map((tool) => (
-            <div
+            <a
               key={tool.name}
+              href={tool.url}
+              target="_blank"
+              rel="noopener noreferrer"
               className="group cursor-pointer"
-              onClick={() => handleRecommendationClick(tool.url)}
             >
               {/* Contenedor de imagen separado */}
               <div className="relative aspect-video bg-zinc-800 rounded mb-3 overflow-hidden border border-[#232323]">
@@ -198,7 +195,7 @@ export default function AIRecommendationsSection({
                   <span>{tool.description}</span>
                 </div>
               </div>
-            </div>
+            </a>
           ))}
         </div>
       </div>
