+++ src/components/Sections/AISecuritySection.tsx
@@ -51,20 +51,13 @@ export default function AISecuritySection({ onViewAll }: AISecuritySectionProps)
 
   const displayedTools = allSecurityTools.slice(0, displayCount);
 
-  const handleToolClick = (url: string) => {
-    window.open(url, '_blank');
-  };
-
   return (
     <div className="py-4 px-4 max-w-7xl mx-auto">
       {/* Contenedor principal con bordes */}
       <div className="bg-zinc-950 rounded-xl p-4 border border-zinc-800">
         {/* Header con título y botón */}
         <div className="flex items-center justify-between mb-8">
-          <h2
-            className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-            style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-          >
+          <h2 className="text-xl md:text-2xl font-semibold text-white text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
             Ciberseguridad
           </h2>
           <button
@@ -79,10 +72,12 @@ export default function AISecuritySection({ onViewAll }: AISecuritySectionProps)
         <div className="md:hidden">
           <div className="space-y-3">
             {displayedTools.map((tool) => (
-              <div
+              <a
                 key={tool.name}
+                href={tool.url}
+                target="_blank"
+                rel="noopener noreferrer"
                 className="flex items-start gap-3 py-1 pl-1 cursor-pointer hover:bg-zinc-900/50 rounded-lg transition-colors p-2"
-                onClick={() => handleToolClick(tool.url)}
               >
                 <Image
                   src={tool.logo}
@@ -96,7 +91,7 @@ export default function AISecuritySection({ onViewAll }: AISecuritySectionProps)
                   <h3 className="font-bold text-white text-base mb-1">{tool.name}</h3>
                   <p className="text-zinc-400 text-sm leading-relaxed">{tool.description}</p>
                 </div>
-              </div>
+              </a>
             ))}
           </div>
         </div>
@@ -104,10 +99,12 @@ export default function AISecuritySection({ onViewAll }: AISecuritySectionProps)
         {/* Vista desktop: Grid de tarjetas con imágenes */}
         <div className="hidden md:grid md:grid-cols-2 lg:grid-cols-4 gap-4">
           {displayedTools.map((tool) => (
-            <div
+            <a
               key={tool.name}
+              href={tool.url}
+              target="_blank"
+              rel="noopener noreferrer"
               className="group cursor-pointer"
-              onClick={() => handleToolClick(tool.url)}
             >
               {/* Contenedor de imagen separado */}
               <div className="relative aspect-video bg-zinc-800 rounded mb-3 overflow-hidden border border-[#232323]">
@@ -132,7 +129,7 @@ export default function AISecuritySection({ onViewAll }: AISecuritySectionProps)
                 <h3 className="font-bold text-white text-base mb-1">{tool.name}</h3>
                 <p className="text-zinc-400 text-sm leading-relaxed">{tool.description}</p>
               </div>
-            </div>
+            </a>
           ))}
         </div>
       </div>
