+++ src/components/Sections/FeaturedToolsSection.tsx
@@ -33,14 +33,11 @@ export default function FeaturedToolsSection() {
           type="application/ld+json"
           dangerouslySetInnerHTML={{ __html: JSON.stringify(itemListJsonLd) }}
         />
-        <h2
-          className="text-xl md:text-2xl font-semibold text-white mb-4 text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]"
-          style={{ fontFamily: 'Inter, system-ui, sans-serif' }}
-        >
+        <h2 className="text-xl md:text-2xl font-semibold text-white mb-4 text-left drop-shadow-[0_0_10px_rgba(255,255,255,0.3)]">
           IAs Destacadas
         </h2>
         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-3 md:gap-4">
-          {featured.map((tool) => (
+          {featured.map((tool, index) => (
             <Link
               key={tool.name}
               href={`/herramienta/${slugify(tool.name)}`}
@@ -53,7 +50,7 @@ export default function FeaturedToolsSection() {
                   fill
                   sizes="(max-width: 640px) 100vw, (max-width: 1024px) 50vw, 25vw"
                   className="w-full h-full object-cover rounded transition-all duration-300 group-hover:scale-105 group-hover:blur-sm"
-                  loading="lazy"
+                  priority={index < 4}
                 />
                 <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-all duration-300 flex items-center justify-center">
                   <div className="bg-zinc-800/90 text-white px-2 md:px-3 py-1 md:py-1.5 rounded-lg text-xs md:text-sm font-medium">
